// common/elements/CLink/interfaces.ts
export interface CLinkProps {
    title?: string;
    default?: true;
    popup?: false;
    tracking?: false;
    event?: 'click';
    linkViaClick?: false;
    checkHashPopUp?: true;
    translate?: true;
    target?: any;
    slug?: any;
    tag?: any;
    children?: any;
    datapopup?: any;
    prefix?: any;
}